Move tests into test directory and add note in README
