+++ test/test_create_store.py
@@ -1,37 +1,7 @@
 import unittest
 import pydux
+from helpers import *
 
-def id(state=[]):
-    def f(result, item):
-        if item['id'] > result:
-            return item['id']
-        else:
-            return result
-    return reduce(f, state, 0) + 1
-
-def reducer_todos(state, action):
-    if state is None:
-        state = []
-    if action['type'] == 'ADD_TODO':
-        copy = state[:]
-        copy.append({'id': id(state), 'text': action['text']})
-        return copy
-    return state
-
-def reducer_todos_reverse(state, action):
-    if state is None:
-        state = []
-    if action['type'] == 'ADD_TODO':
-        copy = state[:]
-        copy.insert(0, ({'id': id(state), 'text': action['text']}))
-        return copy
-    return state
-
-def unknown_action():
-    return {'type': 'UNKNOWN_ACTION'}
-
-def add_todo(text):
-    return {'type': 'ADD_TODO', 'text': text}
 
 class TestCreateStore(unittest.TestCase):
     def test_exposes_public_api(self):
@@ -201,13 +171,3 @@ class TestCreateStore(unittest.TestCase):
                 'text': 'Surely'
             }
         ])
-
-
-
-
-
-
-
-
-if __name__ == '__main__':
-    unittest.main()
